Refactor code
Split up the code across 3 files:
opts - argument parsing
lib - main procfd logic
main - renders results
